% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_control.r
\name{change_control}
\alias{change_control}
\title{configure control plan}
\usage{
change_control(l_config, t_strt, t_stop, control_id, add = FALSE)
}
\arguments{
\item{l_config}{list of config parameters}

\item{t_strt}{vector of start times}

\item{t_stop}{vector of stop times}

\item{control_id}{vector of control_ids for the times}

\item{add}{whether to add the new controls to exisiting ones, default FALSE to replace}
}
\value{
list with modified control_plan
}
\description{
allows start & stop of controls to be set, depends on controls available in config file
}
\examples{
l_config <- read_config()
l_config2 <- change_control(l_config, t_strt=1, t_stop=15, control_id='irs_pyr')
l_config2 <- change_control(l_config, t_strt=c(1,15), t_stop=c(10,20), control_id='irs_pyr')
l_config2 <- change_control(l_config, t_strt=c(1,11), t_stop=c(10,20), control_id=c('irs_pyr','irs_ops'))
l_config2 <- change_control(l_config, t_strt=c(1,11), t_stop=c(10,20), control_id=c('irs_pyr','irs_ddt'))
plot_sim2( run_sim2(l_config=l_config2))
}

