% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_resistance.r
\name{change_resistance}
\alias{change_resistance}
\title{change resistance rate}
\usage{
change_resistance(resistance, resist_incr, resist_decr, resistance_on)
}
\arguments{
\item{resistance}{resistance rate in this timestep}

\item{resist_incr}{increase in resistance in presence of insecticide}

\item{resist_decr}{decrease in resistance in absence of insecticide}

\item{resistance_on}{whether the insecticide applied prompts resistance 0=no, 1=yes}
}
\value{
float resistance in next timestep
}
\description{
change resistance to one insecticide
}
\examples{
change_resistance(resistance=0.2, resist_incr=0.4, resist_decr=0.2, resistance_on=1)
}

