% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config_control_plan.r
\name{config_control_plan}
\alias{config_control_plan}
\title{configure control plan}
\usage{
config_control_plan(l_config, time_strt, time_stop, control_id, add = FALSE)
}
\arguments{
\item{l_config}{list of config parameters}

\item{time_strt}{vector of start times}

\item{time_stop}{vector of stop times}

\item{control_id}{vector of control_ids for the times}

\item{add}{whether to add the new controls to exisiting ones, default FALSE to replace}
}
\value{
list with modified control_plan
}
\description{
allows start & stop of controls to be set, depends on controls available in config file
}
\examples{
l_config <- read_config()
l_config2 <- config_control_plan(l_config, time_strt=1, time_stop=15, control_id='irs_pyr')
l_config2 <- config_control_plan(l_config, time_strt=c(1,15), time_stop=c(10,20), control_id='irs_pyr')
l_config2 <- config_control_plan(l_config, time_strt=c(1,11), time_stop=c(10,20), control_id=c('irs_pyr','irs_ops'))
l_config2 <- config_control_plan(l_config, time_strt=c(1,11), time_stop=c(10,20), control_id=c('irs_pyr','irs_ddt'))
plot_sim2( run_sim2(l_config=l_config2))
}

