% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.r
\name{config_controls}
\alias{config_controls}
\title{configure controls available}
\usage{
config_controls(l_config, control_id, control_name = NA, control_desc = NA,
  vector_id, control_kill_rate, add = FALSE)
}
\arguments{
\item{l_config}{list of config parameters}

\item{control_id}{vector of control_ids}

\item{control_name}{vector of control_names (optional)}

\item{control_desc}{vector of control descriptions (optional)}

\item{vector_id}{vector of vector_ids}

\item{control_kill_rate}{vector of control_kill_rates}

\item{add}{whether to add, default FALSE to replace}
}
\value{
list with modified controls
}
\description{
modifies the input from config files
}
\examples{
l_config <- read_config()
l_config2 <- config_controls(l_config, control_id=c('irs_pyr','irs_ops'), vector_id='an_gamb', control_kill_rate=c(0.3,0.4))
plot_sim( run_sim(l_config=l_config2))
}

