% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/config.r
\name{config_resistances}
\alias{config_resistances}
\title{configure resistances}
\usage{
config_resistances(l_config, resistance_id, resistance_name = NA, control_id,
  resistance_strength, resistance_incr, resistance_decr, add = FALSE)
}
\arguments{
\item{l_config}{list of config parameters}

\item{resistance_id}{vector of resistance_ids}

\item{resistance_name}{vector of resistance_names (optional)}

\item{control_id}{vector of control_ids that each resistance works on}

\item{resistance_strength}{vector of resistance_strengths}

\item{resistance_incr}{vector of resistance increase params}

\item{resistance_decr}{vector of resistance decrease params}

\item{add}{whether to add, default FALSE to replace}
}
\value{
list with modified controls
}
\description{
modifies the input from config files
}
\examples{
l_config <- read_config()
l_config2 <- config_resistances(l_config, resistance_id="target_site", control_id=c('irs_ops'), resistance_strength=c(0.8), resistance_incr=c(0.2), resistance_decr=c(0.1))
plot_sim_oldcc( run_sim_oldcc(l_config=l_config2))
}

