% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expand_season.r
\name{expand_season}
\alias{expand_season}
\title{expand a seasonal pattern string in months to a vector of values by days,weeks or months}
\usage{
expand_season(season_string = "6:0.1;6:0.9", return_tstep = "days")
}
\arguments{
\item{season_string}{a season string of months:value;months:value e.g. "6:0.1;6:0.9"}

\item{return_tstep}{either 'months', 'weeks' or 'days'[default]}
}
\value{
a vector of values by day
}
\description{
initially assumes 28 days per month (4*7 day weeks) 336 day year
}
\examples{
tst <- expand_season(season_string="6:0.1;6:0.9")
}

