% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_sim.r
\name{init_sim}
\alias{init_sim}
\title{initialise simulation results list based on config file}
\usage{
init_sim(num_tsteps = 20, l_config = NULL)
}
\arguments{
\item{num_tsteps}{number of timesteps to put in dataframe}

\item{l_config}{list of config parameters}
}
\value{
list to contain simulation results
}
\description{
create list to contain simulation results
}
\examples{
dF <- init_sim(20, l_config=read_config())
}

