% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/init_sim.r
\name{init_sim_oldest}
\alias{init_sim_oldest}
\title{initialise simulation results dataframe}
\usage{
init_sim_oldest(num_tsteps = 20)
}
\arguments{
\item{num_tsteps}{number of timesteps to put in dataframe}
}
\value{
dataframe to contain simulation results
}
\description{
create dataframe to contain simulation results
}
\examples{
dF <- init_sim_oldest(20)
}

