% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/is_control_incr_resist.r
\name{is_control_incr_resist}
\alias{is_control_incr_resist}
\title{tests whether controls used promote the resistance present}
\usage{
is_control_incr_resist(controls_used, l_config)
}
\arguments{
\item{controls_used}{vector of numbers named with control_ids}

\item{l_config}{list of config parameters}
}
\value{
integer 0=no, 1 =yes
}
\description{
based on controls used and a configuration file
}
\examples{
controls_used <- 1
names(controls_used) <- 'irs_pyr'
is_control_incr_resist( controls_used, read_config())
}

