% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config_gdocs.r
\name{plot_config_gdocs}
\alias{plot_config_gdocs}
\title{plot parts of config files to aid checking}
\usage{
plot_config_gdocs(locations)
}
\arguments{
\item{locations}{dataframe from locations config file}
}
\value{
maybe nothing, produces a plot
}
\description{
plot parts of config files to aid checking
}

