% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_config_gdocs_emergences.r
\name{plot_config_gdocs_emergences}
\alias{plot_config_gdocs_emergences}
\title{plot parts of config files to aid checking}
\usage{
plot_config_gdocs_emergences()
}
\value{
maybe nothing, produces a plot
}
\description{
plot parts of config files to aid checking
}

