% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim_oldest}
\alias{plot_sim_oldest}
\title{plot simulation results}
\usage{
plot_sim_oldest(dF)
}
\arguments{
\item{dF}{dataframe containing simulation results}
}
\value{
maybe nothing, produces a plot
}
\description{
plot insecticide use and resulting changes in population and resistance
}
\examples{
#blank plot
dF <- init_sim_oldest(20)
plot_sim_oldest(dF)
#default run
plot_sim_oldest( run_sim_oldest())
#modify params
plot_sim_oldest( run_sim_oldest( rate_insecticide_kill = 0.3, resist_incr = 0.05 ))
}

