% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim_pop.R
\name{plot_sim_pop}
\alias{plot_sim_pop}
\title{plot simulation pop results for newest flexible config driven simulations}
\usage{
plot_sim_pop(l_time, plot_emergence = TRUE, verbose = FALSE,
  time_label = "weeks", plot_type = "month", subset_data = TRUE,
  leg_pos = "bottomleft", title = NULL)
}
\arguments{
\item{l_time}{list containing simulation results}

\item{plot_emergence}{whether to add emergence rate to population plot}

\item{verbose}{whether to output diagnostics to console}

\item{time_label}{time label to add to x axis, default='weeks'}

\item{plot_type}{'month' to just plot points per month, '6month' to plot 6months}

\item{subset_data}{whether to subset_data, set to FALSE when calling from other plot funcs}

\item{leg_pos}{legend position defaults to 'bottomleft', other combinations of 'top','right' etc.}

\item{title}{plot title added at top}
}
\value{
maybe nothing, produces a plot
}
\description{
pop only (with emergence) not interventions or resistance
}
\examples{
#blank plot
l_time <- init_sim(20)
plot_sim_pop(l_time)
#default run
plot_sim_pop( run_sim())
#modify params
plot_sim_pop( run_sim( insecticide_kill = 0.3, resist_incr = 0.05 ))
}

