% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read_config.r
\name{read_config}
\alias{read_config}
\title{reads in game model configuration files}
\usage{
read_config(in_folder = NULL)
}
\arguments{
\item{in_folder}{folder containing the config files}
}
\value{
list containing configuration details
}
\description{
creates configuration list that can be used to specify game behaviour
}
\examples{
list_config <- read_config()
}

