% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim.r
\name{run_sim}
\alias{run_sim}
\title{run flexible simulation of population and resistance change based on emergence driven by config file}
\usage{
run_sim(num_tsteps = 20, pop_start = 0.5, resist_freq_start = 0.1,
  resist_intensity_start = 1, resist_mech = "metabolic", survival = 0.7,
  emergence = 0.3, rate_insecticide_kill = 0.8, resistance_modifier = 1,
  resist_incr = 0.2, resist_decr = 0.1, l_config = NULL, randomness = 0)
}
\arguments{
\item{num_tsteps}{number of timesteps to run simulation}

\item{pop_start}{start vector population}

\item{resist_freq_start}{frequency of resistance at start, 0 to 1}

\item{resist_intensity_start}{intensity of resistance at start, 1 to 10}

\item{survival}{adult survival rate}

\item{emergence}{emerging adults, can be a vector can be greater than 1}

\item{rate_insecticide_kill}{kill rate due to insecticide}

\item{resistance_modifier}{modifies effect of resistance}

\item{resist_incr}{increase in resistance when correct insecticide present}

\item{resist_decr}{decrease in resistance when correct insecticide absent}

\item{l_config}{list of config parameters}

\item{randomness}{0-1 0=none, 1=maximum}
}
\value{
list of simulation results
}
\description{
some params driven by config file, others by function args
}
\examples{
l_time <- run_sim(pop_start=0.5, resist_freq_start=0.2, survival=0.8, emergence=0.2, rate_insecticide_kill=0.4, resistance_modifier=1)
#plot default run
plot_sim( run_sim())
#modify params
plot_sim( run_sim( rate_insecticide_kill = 0.3, resist_incr = 0.05 ))
#modify config file
l_config <- read_config()
l_config2 <- config_plan(l_config, t_strt=c(1,11), t_stop=c(10,20), control_id=c('irs_pyr','irs_ddt'))
plot_sim( run_sim(l_config=l_config2, resist_incr=0.1))
}

