% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sim_old.r
\name{run_sim_oldcc}
\alias{run_sim_oldcc}
\title{run flexible simulation of population and resistance change driven by config file}
\usage{
run_sim_oldcc(num_tsteps = 20, pop_start = 0.5, resist_freq_start = 0.1,
  rate_growth = 0.2, carry_cap = 1, rate_insecticide_kill = 0.2,
  resistance_modifier = 1, resist_incr = 0.2, resist_decr = 0.1,
  l_config = NULL, randomness = 0, never_go_below = 0.01)
}
\arguments{
\item{num_tsteps}{number of timesteps to run simulation}

\item{pop_start}{start vector population}

\item{resist_freq_start}{effect of resistance on insecticide kill rate}

\item{rate_growth}{population growth rate}

\item{carry_cap}{carrying capacity (K) in the logistic model}

\item{rate_insecticide_kill}{kill rate due to insecticide}

\item{resistance_modifier}{modifies effect of resistance}

\item{resist_incr}{increase in resistance when correct insecticide present}

\item{resist_decr}{decrease in resistance when correct insecticide absent}

\item{l_config}{list of config parameters}

\item{randomness}{0-1 0=none, 1=maximum}

\item{never_go_below}{restock at this level if pop goes below it}
}
\value{
list of simulation results
}
\description{
some params driven by config file, others by function args
}
\examples{
l_time <- run_sim_oldcc(pop_start=0.5, resist_freq_start=0.2, rate_growth=0.4, carry_cap=1, rate_insecticide_kill=0.4, resistance_modifier=1)
#plot default run
plot_sim_oldcc( run_sim_oldcc())
#modify params
plot_sim_oldcc( run_sim_oldcc( rate_insecticide_kill = 0.3, resist_incr = 0.05 ))
#modify config file
l_config <- read_config()
l_config2 <- config_plan(l_config, t_strt=c(1,11), t_stop=c(10,20), control_id=c('irs_pyr','irs_ddt'))
plot_sim_oldcc( run_sim_oldcc(l_config=l_config2, resist_incr=0.1))
}

