% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/run_sim.r
\name{run_sim_oldest}
\alias{run_sim_oldest}
\title{run simulation of population and resistance change}
\usage{
run_sim_oldest(num_tsteps = 20, pop_start = 0.5,
  rate_resistance_start = 0.1, rate_growth = 0.2, carry_cap = 1,
  rate_insecticide_kill = 0.2, resistance_modifier = 1, resist_incr = 0.2,
  resist_decr = 0.1, use_pyr = rep(1, num_tsteps), use_ddt = NA,
  use_ops = NA, use_car = NA, randomness = 0)
}
\arguments{
\item{num_tsteps}{number of timesteps to run simulation}

\item{pop_start}{start vector population}

\item{rate_resistance_start}{effect of resistance on insecticide kill rate}

\item{rate_growth}{population growth rate}

\item{carry_cap}{carrying capacity (K) in the logistic model}

\item{rate_insecticide_kill}{kill rate due to insecticide}

\item{resistance_modifier}{modifies effect of resistance}

\item{resist_incr}{increase in resistance when correct insecticide present}

\item{resist_decr}{decrease in resistance when correct insecticide absent}

\item{use_pyr}{use pyrethroids NA for no, 1 for yes, or a vector e.g. c(NA,1) to give alternate use}

\item{use_ddt}{use ddt see use_pyr}

\item{use_ops}{use organophosphates see use_pyr}

\item{use_car}{use carbamates see use_pyr}

\item{randomness}{0-1 0=none, 1=maximum}
}
\value{
dataframe of simulation results
}
\description{
initially accepts single arg values next get it to accept vectors by time
}
\examples{
dF <- run_sim_oldest(pop_start=0.5, rate_resistance_start=0.2, rate_growth=0.4, carry_cap=1, rate_insecticide_kill=0.4, resistance_modifier=1)
#plot default run
plot_sim_oldest( run_sim_oldest())
#modify params
plot_sim_oldest( run_sim_oldest( rate_insecticide_kill = 0.3, resist_incr = 0.05 ))
#alternate use of pyr
plot_sim_oldest( run_sim_oldest(use_pyr=c(NA,1)))
}

