% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vecbyloc_from_gdocs.R
\name{vecbyloc_from_gdocs}
\alias{vecbyloc_from_gdocs}
\title{create a dataframe of vector behaviour by location from the gdoc config files stored in data}
\usage{
vecbyloc_from_gdocs()
}
\value{
dataframe
}
\description{
returned dataframe has 1 row per vector,location combination
}

