% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_pop.r
\name{change_pop}
\alias{change_pop}
\title{change vector population based on emergence}
\usage{
change_pop(pop, survival, emergence, rate_insecticide_kill, rate_resistance,
  resistance_modifier, resist_intensity = 10, insecticide_on, resistance_on,
  randomness = 0, never_go_below = 0.01, verbose = FALSE)
}
\arguments{
\item{pop}{vector population in this timestep}

\item{survival}{adult survival rate}

\item{emergence}{emerging adults, can be a vector can be greater than 1}

\item{rate_insecticide_kill}{kill rate due to insecticide}

\item{rate_resistance}{effect on resistance on insecticide kill rate}

\item{resistance_modifier}{modifies effect of resistance (unlikely to be used in game)}

\item{resist_intensity}{intensity of resistance 1,2,5 & 10 fold}

\item{insecticide_on}{whether insecticide is applied 0=no, 1=yes}

\item{resistance_on}{whether there is resistance to the applied insecticide 0=no, 1=yes}

\item{randomness}{0-1 0=none, 1=maximum}

\item{never_go_below}{restock at this level if pop goes below it}

\item{verbose}{whether to output diagnostics to console}
}
\value{
float population in next timestep
}
\description{
simple simulation of vector population based upon emergence suggested by Nakul Chitnis
}
\examples{
change_pop(pop=0.5, survival=0.4, emergence=1, rate_insecticide_kill=0.4, rate_resistance=0.2, resistance_modifier=1, resistance_on=1, insecticide_on=1)
}

