% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/change_pop.r
\name{change_pop}
\alias{change_pop}
\title{change vector population}
\usage{
change_pop(pop, rate_growth, carry_cap, rate_insecticide_kill, rate_resistance,
  resistance_modifier, insecticide_on, resistance_on, randomness = 0)
}
\arguments{
\item{pop}{vector population in this timestep}

\item{rate_growth}{population growth rate}

\item{carry_cap}{carrying capacity (K) in the logistic model}

\item{rate_insecticide_kill}{kill rate due to insecticide}

\item{rate_resistance}{effect on resistance on insecticide kill rate}

\item{resistance_modifier}{modifies effect of resistance}

\item{insecticide_on}{whether insecticide is applied 0=no, 1=yes}

\item{resistance_on}{whether there is resistance to the applied insecticide 0=no, 1=yes}

\item{randomness}{0-1 0=none, 1=maximum}
}
\value{
float population in next timestep
}
\description{
simple logistic model to change vector population based on growth rate, K, insecticide and resistance
}
\examples{
change_pop(pop=0.5, rate_growth=0.4, carry_cap=1, rate_insecticide_kill=0.4, rate_resistance=0.2, resistance_modifier=1, resistance_on=1, insecticide_on=1)
}

