% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create_lookup.r
\name{create_lookup}
\alias{create_lookup}
\title{create an example lookup table that could be used by a game to direct it's behaviour}
\usage{
create_lookup(input_values = list(use_pyr = c(0, 1), use_ddt = c(0, 1),
  use_ops = c(0, 1), use_car = c(0, 1), pop_vector = seq(0.1, 1, 0.1),
  resist_pyr = seq(0.1, 1, 0.1)), rate_growth = 0.4,
  rate_insecticide_kill = 0.4, resistance_modifier = 1, carry_cap = 1,
  resist_incr = 0.4, resist_decr = 0.4, write_csv = "demoLookupTable.csv")
}
\arguments{
\item{input_values}{a list of inputs and values}

\item{rate_growth}{vector popn growth rate}

\item{rate_insecticide_kill}{insecticide kill rate}

\item{resistance_modifier}{resistance modifier}

\item{carry_cap}{carrying capacity}

\item{resist_incr}{increase in resistance when correct insecticide present}

\item{resist_decr}{decrease in resistance when correct insecticide absent}

\item{write_csv}{a filename or NULL for no csv output}
}
\value{
a lookup table of input values and population and resistance outputs
}
\description{
this is a prototype to explore what a lookup table might look like
}
\examples{
create_lookup(write_csv=NULL)
}

