% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{plot simulation results for newest flexible config driven simulations}
\usage{
plot_sim(l_time, plot_emergence = TRUE, plot_thresholds = TRUE,
  verbose = FALSE, time_label = "weeks", plot_type = "month_only")
}
\arguments{
\item{l_time}{list containing simulation results}

\item{plot_emergence}{whether to add emergence rate to population plot}

\item{plot_thresholds}{whether to add WHO resistance threholds to resistance plot}

\item{verbose}{whether to output diagnostics to console}

\item{time_label}{time label to add to x axis, default='weeks'}

\item{plot_type}{'month_only' to just plot points per month}
}
\value{
maybe nothing, produces a plot
}
\description{
plot insecticide use and resulting changes in population and resistance
}
\examples{
#blank plot
l_time <- init_sim(20)
plot_sim(l_time)
#default run
plot_sim( run_sim())
#modify params
plot_sim( run_sim( insecticide_kill = 0.3, resist_incr = 0.05 ))
}

