% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_sim.R
\name{plot_sim}
\alias{plot_sim}
\title{plot simulation results}
\usage{
plot_sim(dF)
}
\arguments{
\item{dF}{dataframe containing simulation results}
}
\value{
maybe nothing, produces a plot
}
\description{
plot insecticide use and resulting changes in population and resistance
}
\examples{
#todo need a way of creating example dF from here
#plot_sim(dF)
}

